lorgnette: cast result of base::bits::AlignUp to size_t
libchrome r1023727 (crrev.com/c/3752744) changed base::bits::AlignUp to
be templated, so that it takes size and alignment of the same type T,
and return result as type T. Type T of the usage in lorgnette/manager.cc
will be deduced as int, causing std::max to fail to compare an int and a
size_t.

BUG=b:248960258
TEST=FEATURES=test emerge-hatch lorgnette with new libchrome

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3915688
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Auto-Submit: Grace Cham <hscham@chromium.org>
Tested-by: Grace Cham <hscham@chromium.org>
Commit-Queue: Grace Cham <hscham@chromium.org>

+++ lorgnette/manager.cc
@@ -868,8 +868,9 @@ ScanState Manager::RunScanLoop(brillo::ErrorPtr* error,
   uint32_t last_progress_value = 0;
   size_t rows_written = 0;
   const size_t kMaxBuffer = 1024 * 1024;
-  const size_t buffer_length = std::max(
-      base::bits::AlignUp(params->bytes_per_line, 4 * 1024), kMaxBuffer);
+  const size_t buffer_length = std::max(static_cast<size_t>(base::bits::AlignUp(
+                                            params->bytes_per_line, 4 * 1024)),
+                                        kMaxBuffer);
   std::vector<uint8_t> image_buffer(buffer_length, '\0');
   // The offset within image_buffer to read to. This will be used within the
   // loop for when we've read a partial image line and need to track data that
